Inclusão Teste Integração

+++ ApiCliente.Application/AppServiceCidade.cs
@@ -2,6 +2,7 @@
 using ApiCliente.Application.Interface;
 using ApiCliente.Application.Interface.Mapper;
 using ApiCliente.Core.Interface.Service;
+using ApiCliente.Domain.Entity;
 using ApiCliente.Domain.Validations;
 using System.Collections.Generic;
 
@@ -11,23 +12,24 @@ namespace ApiCliente.Application
     {
         private readonly IServiceCidade _cidadeService;
         private readonly IMapperCidade _cidadeMapper;
-
-
-        public AppServiceCidade(IServiceCidade cidadeService, IMapperCidade cidadeMapper)
+        private readonly IServiceCliente _clienteService;
+        private readonly IMapperCliente _clienteMapper;
+        public AppServiceCidade(IServiceCidade cidadeService, IMapperCidade cidadeMapper, IMapperCliente clienteMapper, IServiceCliente clienteService)
         {
             _cidadeService = cidadeService;
             _cidadeMapper = cidadeMapper;
+            _clienteService = clienteService;
+            _clienteMapper = clienteMapper;
         }
 
         public void Add(CidadeDto obj)
         {
+            //validação Nome da cidade e Estado
             CidadeValidation cidadeValidation = new CidadeValidation();
-
             if (cidadeValidation.ValidarNomeCidade(obj.Nome))
                 throw new System.ArgumentException("O campo Nome está vazio ou tem mais de 20 caracteres", "Erro cliente");
             if (cidadeValidation.ValidarEstado(obj.Estado))
                 throw new System.ArgumentException("O campo Estado está vazio ou tem mais de 2 caracteres (Ex: SP)", "Erro cliente");
-
 
             var objCidade = _cidadeMapper.MapperToEntity(obj);
             _cidadeService.Add(objCidade);
@@ -38,28 +40,28 @@ namespace ApiCliente.Application
             var objCidade = _cidadeService.GetAll();
             return _cidadeMapper.MapperListCidades(objCidade);
         }
-
         public CidadeDto GetById(int id)
         {
             var objCidade = _cidadeService.GetById(id);
             return _cidadeMapper.MapperToDTO(objCidade);
         }
 
-        public void Remove(int id)
+        public void Remove(int id, CidadeDto obj)
         {
             var objCidade = _cidadeService.GetById(id);
-            _cidadeService.Remove(objCidade);
+            var objCidadeId = _clienteService.GetByCidadeId(objCidade.Id);
+            if (objCidadeId == null)
+             _cidadeService.Remove(objCidade);
         }
 
         public void Update(int id, CidadeDto obj)
         {
+            //validação nome da cidade e estado
             CidadeValidation cidadeValidation = new CidadeValidation();
-
             if (cidadeValidation.ValidarNomeCidade(obj.Nome))
-                throw new System.ArgumentException("Campo logradouro é obrigatório ou tem mais de 20 caracteres", "Erro cliente");
+                throw new System.ArgumentException("Campo logradouro é obrigatório ou tem mais de 20 caracteres", "Erro cidade");
             if (cidadeValidation.ValidarEstado(obj.Estado))
-                throw new System.ArgumentException("O campo estado é obrigatório ou tem mais de 2 caracteres", "Erro cliente");
-
+                throw new System.ArgumentException("O campo estado é obrigatório ou tem mais de 2 caracteres", "Erro cidade");
 
             var objCidade = _cidadeService.GetById(id);
